7 authorization

+++ backend/modules/server/handlers.go
@@ -64,7 +64,13 @@ func (h *Handler) Create(c *gin.Context) {
 	}
 
 	// insert in user the new server created
-	user.Servers = append(user.Servers, server.UniqueName)
+	if user.Servers == nil {
+		user.Servers = map[string]string{
+			server.UniqueName: "owner",
+		}
+	} else {
+		user.Servers[server.UniqueName] = "owner"
+	}
 	err = h.RepoUser.Update(user)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to update user"})
@@ -124,6 +130,12 @@ func (h *Handler) Update(c *gin.Context) {
 		return
 	}
 
+	// role check
+	if utils.IsRoleAtLeastAdmin(c, server.UniqueName) {
+		c.Status(http.StatusUnauthorized)
+		return
+	}
+
 	// update data
 	server.Name = rb.Name
 	server.Image = rb.Image
@@ -147,14 +159,26 @@ func (h *Handler) Delete(c *gin.Context) {
 		return
 	}
 
+	server, err := h.Repo.Read(objectId)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to retreive server"})
+		return
+	}
+
+	// role check
+	if utils.IsRoleAdmin(c, server.UniqueName) {
+		c.Status(http.StatusUnauthorized)
+		return
+	}
+
 	// try deleting
 	isDeleted, err := h.Repo.Delete(objectId)
 	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to delete user"})
+		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to delete server"})
 		return
 	}
 	if !isDeleted {
-		c.JSON(http.StatusNotFound, gin.H{"error": "Failed to retreive user"})
+		c.JSON(http.StatusNotFound, gin.H{"error": "Failed to retreive server"})
 		return
 	}
 
@@ -175,6 +199,12 @@ func (h *Handler) Invite(c *gin.Context) {
 	server, err := h.Repo.Read(objectId)
 	if err != nil {
 		c.JSON(http.StatusNotFound, gin.H{"error": "Failed to retreive server"})
+		return
+	}
+
+	// role check
+	if utils.IsRoleAtLeastMember(c, server.UniqueName) {
+		c.Status(http.StatusUnauthorized)
 		return
 	}
 
@@ -227,14 +257,22 @@ func (h *Handler) Join(c *gin.Context) {
 	}
 
 	// check already in the list
-	if utils.Contains(server.Users, user.UniqueName) || utils.Contains(user.Servers, server.UniqueName) {
+	_, existsInServer := server.Users[user.UniqueName]
+
+	if existsInServer {
 		c.JSON(http.StatusBadRequest, gin.H{"error": "User already joined server"})
 		return
 	}
 
 	// todo: need to take user from access token
-	server.Users = append(server.Users, user.UniqueName)
-	user.Servers = append(user.Servers, server.UniqueName)
+	server.Users[user.UniqueName] = "member"
+	if user.Servers == nil {
+		user.Servers = map[string]string{
+			server.UniqueName: "member",
+		}
+	} else {
+		user.Servers[server.UniqueName] = "owner"
+	}
 
 	err = h.Repo.Update(server)
 	if err != nil {
@@ -287,14 +325,16 @@ func (h *Handler) Leave(c *gin.Context) {
 	}
 
 	// check already in the list
-	if !utils.Contains(server.Users, user.UniqueName) && !utils.Contains(user.Servers, server.UniqueName) {
+	_, existsInServer := server.Users[user.UniqueName]
+
+	if !existsInServer {
 		c.JSON(http.StatusBadRequest, gin.H{"error": "User not in the server"})
 		return
 	}
 
 	// todo: need to take user from access token
-	server.Users = utils.Remove(server.Users, user.UniqueName)
-	user.Servers = utils.Remove(user.Servers, server.UniqueName)
+	delete(server.Users, user.UniqueName)
+	delete(user.Servers, server.UniqueName)
 
 	err = h.Repo.Update(server)
 	if err != nil {
